requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
pycryptodome
websockets
pydantic
